+++ source/databricks/settlement_report/settlement_report_job/wholesale/database_definitions.py
@@ -25,4 +25,4 @@ class WholesaleResultsDatabase:
     LATEST_CALCULATIONS_BY_DAY_VIEW_NAME = "latest_calculations_by_day_v1"
     ENERGY_V1_VIEW_NAME = "energy_v1"
     ENERGY_PER_ES_V1_VIEW_NAME = "energy_per_es_v1"
-    WHOLESALE_RESULTS_VIEW_NAME = "amounts_per_charge_v1"  # for some reason we call amounts per charge for wholesale results
+    AMOUNTS_PER_CHARGE_VIEW_NAME = "amounts_per_charge_v1"  # for some reason we call amounts per charge for wholesale results
