+++ source/databricks/settlement_report/tests/test_factories/default_test_data_spec.py
@@ -20,7 +20,7 @@ from test_factories.metering_point_time_series_factory import (
     MeteringPointTimeSeriesTestDataSpec,
 )
 from test_factories.energy_factory import EnergyTestDataSpec
-from test_factories.wholesale_factory import WholesaleTestDataSpec
+from test_factories.amounts_per_charge_factory import AmountsPerChargeTestDataSpec
 
 DEFAULT_FROM_DATE = datetime(2024, 1, 1, 23)
 DEFAULT_TO_DATE = DEFAULT_FROM_DATE + timedelta(days=1)
@@ -131,7 +131,7 @@ def create_time_series_data_spec(
     )
 
 
-def create_wholesale_data_spec(
+def create_amounts_per_charge_data_spec(
     calculation_id: str = DEFAULT_CALCULATION_ID,
     calculation_type: CalculationTypeDataProductValue = CalculationTypeDataProductValue.WHOLESALE_FIXING,
     calculation_version: int = DEFAULT_CALCULATION_VERSION,
@@ -151,8 +151,8 @@ def create_wholesale_data_spec(
     quantity_qualities: list[str] = ["measured"],
     price: Decimal = Decimal("0.005"),
     amount: Decimal = Decimal("0.005"),
-) -> WholesaleTestDataSpec:
-    return WholesaleTestDataSpec(
+) -> AmountsPerChargeTestDataSpec:
+    return AmountsPerChargeTestDataSpec(
         calculation_id=calculation_id,
         calculation_type=calculation_type,
         calculation_version=calculation_version,
